+++ supabase-auth/src/commonMain/kotlin/io/github/androidpoet/supabase/auth/AuthClientImpl.kt
@@ -763,7 +763,6 @@ internal class AuthClientImpl(
         val body =
             defaultJson.encodeToString(
                 MfaVerifyRequest(
-                    factorId = factorId,
                     challengeId = challengeId,
                     code = code,
                     webauthn = webauthn,